Personal Library Manager/requirements.txt
streamlit
pandas
Pillow
matplotlib